chore: separate tfm case amendments e2e tests for shorter execution time
