+++ e2e-tests/tfm/cypress/e2e/journeys/case-amendments/5. facility-value.spec.js
@@ -1,10 +1,10 @@
-import relative from '../../../relativeURL';
-import facilityPage from '../../../pages/facilityPage';
-import amendmentsPage from '../../../pages/amendments/amendmentsPage';
-import MOCK_DEAL_AIN from '../../../../fixtures/deal-AIN';
-import dateConstants from '../../../../../../e2e-fixtures/dateConstants';
-import { PIM_USER_1 } from '../../../../../../e2e-fixtures';
-import { MOCK_MAKER_TFM, ADMIN_LOGIN } from '../../../../fixtures/users-portal';
+import relative from '../../relativeURL';
+import facilityPage from '../../pages/facilityPage';
+import amendmentsPage from '../../pages/amendments/amendmentsPage';
+import MOCK_DEAL_AIN from '../../../fixtures/deal-AIN';
+import dateConstants from '../../../../../e2e-fixtures/dateConstants';
+import { PIM_USER_1 } from '../../../../../e2e-fixtures';
+import { MOCK_MAKER_TFM, ADMIN_LOGIN } from '../../../fixtures/users-portal';
 
 context('Amendments - Facility value', () => {
   let dealId;
