+++ e2e-tests/tfm/cypress/e2e/journeys/case-amendments/add-bank-decision-underwriter-declined.spec.js
@@ -1,14 +1,14 @@
-import relative from '../../../relativeURL';
-import caseSubNavigation from '../../../partials/caseSubNavigation';
-import facilityPage from '../../../pages/facilityPage';
-import amendmentsPage from '../../../pages/amendments/amendmentsPage';
-import caseDealPage from '../../../pages/caseDealPage';
-import MOCK_DEAL_AIN from '../../../../fixtures/deal-AIN';
-import dateConstants from '../../../../../../e2e-fixtures/dateConstants';
-import { PIM_USER_1, UNDERWRITER_MANAGER_1, UNDERWRITER_MANAGER_DECISIONS } from '../../../../../../e2e-fixtures';
-import { MOCK_MAKER_TFM, ADMIN_LOGIN } from '../../../../fixtures/users-portal';
-import pages from '../../../pages';
-import { NOT_ADDED } from '../../../../fixtures/constants';
+import relative from '../../relativeURL';
+import caseSubNavigation from '../../partials/caseSubNavigation';
+import facilityPage from '../../pages/facilityPage';
+import amendmentsPage from '../../pages/amendments/amendmentsPage';
+import caseDealPage from '../../pages/caseDealPage';
+import MOCK_DEAL_AIN from '../../../fixtures/deal-AIN';
+import dateConstants from '../../../../../e2e-fixtures/dateConstants';
+import { PIM_USER_1, UNDERWRITER_MANAGER_1, UNDERWRITER_MANAGER_DECISIONS } from '../../../../../e2e-fixtures';
+import { MOCK_MAKER_TFM, ADMIN_LOGIN } from '../../../fixtures/users-portal';
+import pages from '../../pages';
+import { NOT_ADDED } from '../../../fixtures/constants';
 
 context('Amendments underwriting - add banks decision - declined by underwriter', () => {
   let dealId;
